moved data into top level folder
