+++ app/components/Projects.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router';
 
 import Cell from './Projects/Cell';
-import data from './Projects/data/projects';
+import data from '../data/projects';
 
 // TODO: Put projects side by side (2x2) instead of (1x4)
